move javalabRedux
