+++ apps/src/javalab/CommitDialog.jsx
@@ -8,7 +8,7 @@ import StylizedBaseDialog, {
 import {connect} from 'react-redux';
 import _ from 'lodash';
 import CommitDialogBody from './CommitDialogBody';
-import {setCommitSaveStatus} from '@cdo/apps/javalab/javalabRedux';
+import {setCommitSaveStatus} from '@cdo/apps/javalab/redux/javalabRedux';
 import {CompileStatus} from './constants';
 import {BackpackAPIContext} from './BackpackAPIContext';
