+++ apps/test/unit/javalab/BackpackTest.js
@@ -2,7 +2,7 @@ import React from 'react';
 import {expect, assert} from '../../util/reconfiguredChai';
 import {mount} from 'enzyme';
 import {registerReducers, stubRedux, restoreRedux} from '@cdo/apps/redux';
-import javalab from '@cdo/apps/javalab/javalabRedux';
+import javalab from '@cdo/apps/javalab/redux/javalabRedux';
 import BackpackClientApi from '@cdo/apps/code-studio/components/backpack/BackpackClientApi';
 import sinon from 'sinon';
 import {UnconnectedBackpack as Backpack} from '@cdo/apps/javalab/Backpack';
