+++ apps/test/unit/javalab/JavalabEditorDialogManagerTest.js
@@ -23,7 +23,7 @@ import VersionHistoryWithCommitsDialog from '@cdo/apps/templates/VersionHistoryW
 import javalabEditor, {
   setAllSourcesAndFileMetadata
 } from '@cdo/apps/javalab/redux/editorRedux';
-import javalab from '@cdo/apps/javalab/javalabRedux';
+import javalab from '@cdo/apps/javalab/redux/javalabRedux';
 import javalabView from '@cdo/apps/javalab/redux/viewRedux';
 
 describe('JavalabEditorDialogManager', () => {
